requirements.txt
asgiref==3.4.1
Django==3.2.6
djangorestframework==3.12.4
mysqlclient @ file:///C:/Users/ACEFR/Desktop/TASK%20USER%20PROFILE/USER_PROFILE/mysqlclient-1.4.6-cp38-cp38-win32.whl
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
